requirements.txt
streamlit==1.31.0
pinecone-client==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
numpy==1.26.3
python-dotenv==1.0.0